feat(Synthesis): Create recaps for teacher and player
