+++ packages/client/src/modules/play/Components/Synthesis/Synthesis.common.styles.tsx
@@ -1,0 +1,11 @@
+import { styled } from "@mui/material";
+import { Card } from "../../../common/components/Card";
+
+export { CardStyled };
+
+const CardStyled = styled(Card)(({ theme }) => ({
+  display: "flex",
+  flexDirection: "column",
+  gap: theme.spacing(2),
+  padding: theme.spacing(2),
+}));
