+++ packages/client/src/modules/play/Components/Synthesis/SynthesisBudget/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisBudget";
