+++ packages/client/src/modules/play/Components/Synthesis/SynthesisCarbon/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisCarbon";
