+++ packages/client/src/modules/play/Components/Synthesis/SynthesisConsumption/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisConsumption";
