+++ packages/client/src/modules/play/Components/Synthesis/SynthesisProduction/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisProduction";
