+++ packages/client/src/modules/play/Components/Synthesis/SynthesisRecapForPlayer/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisRecapForPlayer";
