+++ packages/client/src/modules/play/Components/Synthesis/SynthesisRecapForTeacher/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SynthesisRecapForTeacher";
