--- packages/client/src/modules/play/Components/Synthesis/index.ts
@@ -1,1 +1,0 @@
-export { SynthesisRecap } from "./Synthesis";
